requirements.txt
# === Core Framework ===
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0

# === EPUB Processing ===
ebooklib>=0.18
html2text>=2024.2.26
beautifulsoup4>=4.12.0
lxml>=5.2.0

# === Vector Store ===
chromadb>=0.5.0

# === Embedding & Reranking (HuggingFace) ===
sentence-transformers>=3.0.0
transformers>=4.44.0
torch>=2.3.0
# For Apple Silicon Mac use: pip install torch torchvision torchaudio

# === OpenAI ===
openai>=1.40.0

# === Evaluation ===
ragas>=0.1.21
datasets>=2.20.0

# === Utilities ===
python-dotenv>=1.0.0
pydantic>=2.7.0
tqdm>=4.66.0
pandas>=2.2.0
numpy>=1.26.0
rich>=13.7.0      # pretty terminal output
tiktoken>=0.7.0   # token counting
